+++ lib/hashie.rb
@@ -31,6 +31,10 @@ module Hashie
     autoload :RubyVersion,       'hashie/extensions/ruby_version'
     autoload :RubyVersionCheck,  'hashie/extensions/ruby_version_check'
 
+    module Coercion
+      autoload :HashieTypes,     'hashie/extensions/coercion/hashie_types'
+    end
+
     module Parsers
       autoload :YamlErbParser, 'hashie/extensions/parsers/yaml_erb_parser'
     end
